+++ src/Form/InvoiceFilterType.php
@@ -50,6 +50,17 @@ class InvoiceFilterType extends AbstractType
                 ],
                 'attr' => ['class' => 'form-element'],
             ])
+            ->add('noCost', ChoiceType::class, [
+                'required' => true,
+                'label' => 'invoices.list_no_cost',
+                'label_attr' => ['class' => 'label'],
+                'choices' => [
+                    '' => null,
+                    'invoices.list_no_cost_true' => true,
+                    'invoices.list_no_cost_false' => false,
+                ],
+                'attr' => ['class' => 'form-element'],
+            ])
         ;
     }