+++ src/Model/Invoices/ConfirmData.php
@@ -2,7 +2,11 @@
 
 namespace App\Model\Invoices;
 
-class ConfirmData
+final class ConfirmData
 {
-    public bool $confirmed = false;
+    public const INVOICE_RECORD_NO = 'INVOICE_RECORD_NO';
+    public const INVOICE_RECORD_YES = 'INVOICE_RECORD_YES';
+    public const INVOICE_RECORD_YES_NO_COST = 'INVOICE_RECORD_YES_NO_COST';
+
+    public string $confirmation = self::INVOICE_RECORD_NO;
 }