+++ assets/fontawesome.js
@@ -4,7 +4,13 @@ import { library, dom } from '@fortawesome/fontawesome-svg-core'
 // Import the icons from the free solid package.
 import {
     faMaximize,
+    faEyeSlash,
+    faMinimize,
 } from '@fortawesome/free-solid-svg-icons'
 
-library.add(faMaximize);
+library.add(
+    faMaximize,
+    faEyeSlash,
+    faMinimize,
+);
 dom.i2svg()