+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   1258: Clean up planning view ui and add scroll to active sprint.
 * [PR-112](https://github.com/itk-dev/economics/pull/112)
   1280: Simplified planning form. Added default value.
+* [PR-113](https://github.com/itk-dev/economics/pull/113)
+  Worklog period filter
 * [PR-110](https://github.com/itk-dev/economics/pull/110)
   1209: No cost invoices
 * [PR-111](https://github.com/itk-dev/economics/pull/111)